+++ include/queue.h
@@ -8,11 +8,10 @@
 #define QUEUE_GET(q)        (q.buf[q.head])
 #define QUEUE_SET(q, val)   (q.buf[q.tail] = val)
 
-struct runqueue {
-    // int head;
-    // int tail;
-    // unsigned long buf[TASK_CAPACITY];
-};
+// struct runqueue {
+//     int head;
+//     int tail;
+//     unsigned long buf[TASK_CAPACITY];
+// };
 
 // struct runqueue RunQueue;
-
