+++ include/syscall.h
@@ -1,0 +1,7 @@
+#define SYS_CORE_TIMER_IRQ_ENABLE   0
+#define SYS_GET_TASKID              1
+#define SYS_EXIT                    2
+#define SYS_FORK                    3    
+#define SYS_UART_READ               4
+#define SYS_UART_WRITE              5   
+#define SYS_KILL                    6
