+++ src/task.c
@@ -2,12 +2,14 @@
 #include "../include/time.h"
 #include "../include/task.h"
 #include "../include/queue.h"
+#include "../include/syscall.h"
 
-struct runqueue RunQueue;
 
 #define INIT_TASK \
-/*cpu_context*/	{ {0,0,0,0,0,0,0,0,0,0,0,0,0}, \
-/* state etc */	0,0,0,0,0 \
+{ \
+/* cpu_context */   {0,0,0,0,0,0,0,0,0,0,0,0,0}, \
+/* user_context */   {0,0,0}, \
+/* state etc */	    0,0,0,0,0 \
 }
 
 struct task init_task = INIT_TASK;
@@ -25,6 +27,7 @@ void task_manager_init()
 {
     // uart_memset(TaskManager.kstack_pool, '\0', 64*4096);
     set_current(&(init_task));
+    TaskManager.queue_bitmap = 0;
     TaskManager.task_num = 0;
 }
 
@@ -40,15 +43,20 @@ void privilege_task_create(void(*func)())
     // uart_hex((unsigned long) func);
     new_task->cpu_context.pc = (unsigned long)ret_from_fork;
     new_task->cpu_context.sp = (unsigned long) &TaskManager.kstack_pool[task_id];
+
+    new_task->user_context.sp_el0 = &TaskManager.ustack_pool[task_id];
+    new_task->user_context.spsr_el1 = 0;
+    new_task->user_context.elr_el1 = 0;
+    
     new_task->task_id = task_id;
     new_task->reschedule_flag = 0;
+    new_task->state = RUN_IN_KERNEL_MODE;
     TaskManager.task_num++;
 
     // put the task into the runqueue
     // QUEUE_SET(RunQueue, (unsigned long) new_task);
     // QUEUE_PUSH(RunQueue);
 }
-
 
 void context_switch(struct task* next)
 {
@@ -110,8 +118,16 @@ void idle()
     }
 }
 
-
-
+/* 
+ ** user porcess for test
+ */
+void user_test()
+{
+    do_exec(foo);
+}
+/*
+ ** takes a function pointer to user code
+ */
 void do_exec(void(*func)())
 {
     // be triggered at the first time execute
@@ -122,24 +138,42 @@ void do_exec(void(*func)())
 
     // _setup_user_content(func);
 
-    unsigned long user_stack = &TaskManager.ustack_pool[current->task_id];
-    unsigned long zero_reg = 0x0;
-    asm volatile("msr sp_el0, %0" :: "r" (user_stack));
-    asm volatile("msr spsr_el1, %0" :: "r" (zero_reg));
-    asm volatile("msr elr_el1, %0" :: "r" (func));
     if (current->reschedule_flag == 1) {
         uart_puts("kernel routine reschedule...\n");
         current->reschedule_flag = 0;
         schedule();
     }
-    asm volatile("eret");
+    if (current->state == RUN_IN_KERNEL_MODE) {
+        // return to user mode
+        current->state = RUN_IN_USER_MODE;
+        unsigned long user_stack = current->user_context.sp_el0;
+        unsigned long user_cpu_state = 0x0;
+        asm volatile("msr sp_el0, %0" :: "r" (user_stack));
+        asm volatile("msr spsr_el1, %0" :: "r" (user_cpu_state));
+        asm volatile("msr elr_el1, %0" :: "r" (func));
+        asm volatile("eret");
+    } else {
+        current->state = RUN_IN_USER_MODE;
+        unsigned long user_stack = current->user_context.sp_el0;
+        unsigned long user_cpu_state = current->user_context.spsr_el1;
+        unsigned long user_pc = current->user_context.elr_el1;
+        asm volatile("msr sp_el0, %0" :: "r" (user_stack));
+        asm volatile("msr spsr_el1, %0" :: "r" (user_cpu_state));
+        asm volatile("msr elr_el1, %0" :: "r" (user_pc));
+        asm volatile("eret");
+    }
 }
 
 void foo() // in el0
 {
+    unsigned long foo_count = 0;
     while(1) {
-        uart_puts("foo\n");        
-        wait_cycles(100000000);
+        uart_puts("foo: ");
+        uart_hex(foo_count++);
+        uart_puts("\n");
+        // syscall - get task_id
+        // asm volatile("mov x8, #1\n" "svc #0\n");        
+        wait_cycles(10000000);
     }
 }
